% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_labels.r
\name{compute_labels}
\alias{compute_labels}
\title{Clustering matching}
\usage{
compute_labels(idc, idc1)
}
\arguments{
\item{idc}{A character vector of reference group labels.}

\item{idc1}{Vector of labels for clusters.}
}
\value{
Vector of corresponding reference labels for clusters.
}
\description{
This function identifies true label groups between reference groups and clusters
}
\examples{

}
