% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_markers.r
\name{cut_markers}
\alias{cut_markers}
\title{Select the top n cluster markers}
\usage{
cut_markers(clusters, markers, ntop)
}
\arguments{
\item{clusters}{A vector of cluster labels.}

\item{markers}{A data.frame as in the output of the FindAllMarkers Seurat function.}

\item{ntop}{The number of top markers you want.}
}
\value{
A list of ntop markers per cluster.
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{

}
