% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defac.R
\name{defac_matrix}
\alias{defac_matrix}
\title{Markers and differentially expressed genes of the simulated data set}
\usage{
defac_matrix(sim)
}
\arguments{
\item{sim}{A Splatter \code{sim} object with simulated cellular groups or paths.}
}
\value{
A list with four outputs:
\itemize{
\item a data frame (de) indicating differentially expressed (DE) genes (1) and
not (0). Columns represent groups and rows are genes.
\item The list with sets of de genes per group (\code{gr_de}).
\item The list with sets of group markers (\code{markers}).
\item The list with sets of markers specific only in one group (\code{sp_markers}).
}
}
\description{
This function gives you information about differentially expressed genes,
markers and and markers specific of only one group. The definition is based
according to the DEFac value provided by the Splatter simulation.
}
\details{
Here a gene is defined marker of a group if its DEFac>1 in that group.
}
\examples{
# TODO
}
