% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_dnn_classify.R
\name{ds_dnn_classify}
\alias{ds_dnn_classify}
\title{This function trains, fits and evaluates a deep neural network (DNN) model.}
\usage{
ds_dnn_classify(dnn_model, model.data, query.data)
}
\arguments{
\item{dnn_model}{The DNN model output of the ds_dnn_model function.}

\item{model_data}{The output of the ds_split_data_dnn function.}

\item{query_data}{The scaled/normalized gene expression data you want to annotate.}
}
\value{
The classification output of the DNN model.
}
\description{
This function creates a DNN model from the reference dataset (scRNA-seq) by using the keras package.
}
\examples{
#not run
enc <- ds_encoder(data = scaled,genes = gg,dims = 2000,verbose = T,hnodes = c(5000))
features <- ds_get_features(enc = enc,data = scaled,genes=gg)
out <- ds_split_data_encoder(features = features,clus = cluster,prop = 0.8,verbose = T)
model <- ds_dnn_model(out = out,hnodes = c(1000),verbose = T,epochs = 10,batch_size = 32)
}
