% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_encoder.R
\name{ds_encoder}
\alias{ds_encoder}
\title{Creation of an autoencoder model}
\usage{
ds_encoder(data, genes, dims, hnodes, verbose = T, activation = "tanh", ...)
}
\arguments{
\item{data}{A scaled matrix of gene expressions like in the \code{scale.data}
of the Seurat object. Rows are genes and columns are cells from the reference
dataset.}

\item{genes}{The set of genes you want to use before the dimensionality reduction. Usually, the genes are those in common with the query dataset.}

\item{dims}{The dimensions of the space you want to reduce the original data.}

\item{hnodes}{The number of the nodes of the intermediate layers of the encoder. The first and last layers are not included.}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}

\item{activation}{Keras parameter for the layer_dense function (Default= "tanh").}

\item{...}{Keras parameters for the layer_dense function.}
}
\value{
The encoder model to reduce the dimensions of the input data.
}
\description{
This function creates an autoencoder from the reference dataset (scRNA-seq)
}
\examples{
#not run
#enc <- encoder(data = scaled,genes = gg,dims = 2000,verbose = T,hnodes = c(5000))
}
