% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_get_features.R
\name{ds_get_features}
\alias{ds_get_features}
\title{Feature extraction via an autoencoder model}
\usage{
ds_get_features(enc, data, genes)
}
\arguments{
\item{enc}{The encoder model output of the ds_encoder function.}

\item{data}{A scaled matrix of gene expressions like in the \code{scale.data}
of the Seurat object. Rows are genes and columns are cells from the reference
dataset.}

\item{genes}{The set of genes you want to use before the dimensionality reduction. Usually, the genes are those in common with the query dataset.}
}
\value{
The features extracted by the encoder.
}
\description{
This function extracts features from the reference dataset (scRNA-seq)
}
\examples{
#not run
enc <- ds_encoder(data = scaled,genes = gg,dims = 2000,verbose = T,hnodes = c(5000))
features <- ds_get_features(enc = enc,data = scaled,genes=gg)
}
