% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_split_data_dnn.R
\name{ds_split_data_dnn}
\alias{ds_split_data_dnn}
\title{This function split the reference data in training and test set to use it as input for the ds_dnn_model function.}
\usage{
ds_split_data_dnn(scale.data, clus, genes, prop = NULL, verbose = TRUE)
}
\arguments{
\item{scale.data}{scale.data A scaled/normalized matrix of gene expressions like in the \code{scale.data}
of the Seurat object. Rows are genes and columns are cells from the reference
dataset.}

\item{clus}{A named factor with the cell type annotations from the reference data.}

\item{genes}{The set of genes you want to use.}

\item{prop}{The proportion of cells for the training dataset (default=0.5).}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}
}
\value{
A list with the following data: train_x= training_data,train_y= categorical cell type variables,
test_x =test_data,test_y=categorical celltype from the test data, classes=original cell types).
}
\description{
This function split the reference data in training and test set to use it as input for the ds_dnn_model function.
}
\examples{
#not run
}
