% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_split_data_encoder.R
\name{ds_split_data_encoder}
\alias{ds_split_data_encoder}
\title{This function split the encoded data in training and test set.}
\usage{
ds_split_data_encoder(features, clus, prop = NULL, verbose = TRUE)
}
\arguments{
\item{features}{The encoder model output of the ds_encoder function.}

\item{clus}{A named factor with the cell type annotations from the reference data.}

\item{prop}{The proportion of cells for the training dataset (default=0.5).}
}
\value{
A list with the following data: train_x= training_data,train_y= categorical cell type variables,
test_x =test_data,test_y=categorical celltype from the test data, classes=original cell types).
}
\description{
This function extracts features from the reference dataset (scRNA-seq)
}
\examples{
#not run
enc <- ds_encoder(data = scaled,genes = gg,dims = 2000,verbose = T,hnodes = c(5000))
features <- ds_get_features(enc = enc,data = scaled,genes=gg)
out <- ds_split_data_encoder(features = features,clus = cluster,prop = 0.8,verbose = T)
}
