% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_split_data_stack.mod.R
\name{ds_split_stack.mod}
\alias{ds_split_stack.mod}
\title{This function split the reference data in validation and test set to use it as input for the ds_stacking_model function.
In this way the joint model will be fitted with data that are different from the original training set.}
\usage{
ds_split_stack.mod(out, clus, prop = NULL, verbose = TRUE)
}
\arguments{
\item{out}{The original output of the ds_split_data_dnn function for one of the model}

\item{clus}{A named factor with the cell type annotations from the reference data.}

\item{prop}{The proportion of cells for the training dataset (default=0.5).}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}
}
\value{
A list with the following data: val_x= validation_data,val_y= categorical cell type variables,
test_x =test_data,test_y=categorical celltype from the test data, classes=original cell types).
}
\description{
This function split the reference data in validation and test set to use it as input for the ds_stacking_model function.
In this way the joint model will be fitted with data that are different from the original training set.
}
\examples{
#not run
}
