% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_stacking_model.R
\name{ds_stacking_model}
\alias{ds_stacking_model}
\title{This function creates a stacking ensemble of 2 or more DNN models.}
\usage{
ds_stacking_model(
  list.models,
  hnodes,
  activation = "relu",
  add_dropout = F,
  pct_dropout = 0.2,
  verbose = T,
  ...
)
}
\arguments{
\item{list.models}{A named list with the different models. The names of the models must be different and correspond to the name given to the name_mod in the ds_dnn_model function.}

\item{hnodes}{The number of the nodes of the intermediate layers of the encoder. The first and last layers are not included.}

\item{activation}{Keras parameter for the layer_dense function (Default= "relu").}

\item{add_dropout}{Keras parameter for the layer_dense function (Default= FALSE).}

\item{pct_dropout}{Keras parameter for the layer_dense function (Default= 0.2).}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}

\item{...}{Other keras parameters.}
}
\value{
The stacked model.
}
\description{
This function creates a DNN model from the reference dataset (scRNA-seq) by using the keras package.
}
\examples{
#not run
stack.mod <- ds_stacking_model(list.models,hnodes = c(1000),activation = "relu",verbose = T)
}
