% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_annotation.R
\name{GOannotation}
\alias{GOannotation}
\title{Compute the weighted GO enrichment of cluster-specific markers.}
\usage{
GOannotation(
  markers,
  go.db,
  species = "mouse",
  ontology.type = "BP",
  reformat.gene.names = FALSE,
  go.score.class = "weight01Score",
  p.val.threshold = 0.05,
  dag.file.prefix = FALSE,
  ngenes = 20
)
}
\arguments{
\item{markers}{A data.frame of cluster specific gene markers like in the
Seurat output of the function FindAllMarkers. The used columns are
\code{c("p_val","cluster","gene")}}

\item{go.db}{Human (\code{org.Hs.eg.db}) or mouse (\code{org.Mm.eg.db}) GO database.
Corresponding R packages have to be installed.}

\item{species}{character indicating the species. Only 'human' or 'mouse'.}

\item{ontology.type}{Ontology family to be examined ("BP"= 'Biological Process',
"MF"='Molecular Function',"CC"='Cellular Component')}

\item{reformat.gene.names}{TODO}

\item{go.score.class}{TODO}

\item{p.val.threshold}{TODO}

\item{dag.file.prefix}{TODO}

\item{ngenes}{TODO}
}
\value{
A list with two elements:
\itemize{
\item \code{GOenrich} contains data.frames with GO enrichments for each cluster.
\item \code{genes_byGO} provides the set of genes from each resulting GO pathway that
are in overlap between your data and the pathway.
}
}
\description{
This function computes the weighted GO enrichment across computational gene
markers by using their p-values.
}
\examples{
# TODO
}
