% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_map.R
\name{identity_map}
\alias{identity_map}
\title{Cell projections onto the reference dataset}
\usage{
identity_map(scale.data, model, gene_cl.ref, p.threshold = NULL)
}
\arguments{
\item{scale.data}{A scaled matrix of gene expressions like in the \code{scale.data}
of the Seurat object. Rows are genes and columns are cells from the reference
dataset.}

\item{model}{The model as in output of the \code{train_model} function.}

\item{gene_cl.ref}{A list of cluster-specific markers. Each element of the list
contains markers of a cell type. Usually only top100 ranked markers are used.}

\item{p.threshold}{Probability threshold to consider a cell classified.
Default=0.5.}
}
\value{
A list with a vector with cell identities predicted by the model (ids)
and a dataframe with probabilities at each identity class.
}
\description{
This function infers cell identities by using the model learned by the
reference dataset.
}
\examples{
# TODO
}
