% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_by_specificity.r
\name{markers_by_specificity}
\alias{markers_by_specificity}
\title{Simulated group markers}
\usage{
markers_by_specificity(rank_df, spec, n_clust)
}
\arguments{
\item{rank_df}{The data frame with the ranked group genes as returned by the rank_sim function.}

\item{spec}{The proportion of top ranked genes. It has to be a number between 0 and 1.}

\item{n_clust}{The number of simulated cell groups.}
}
\value{
A list with n_clust elements representing their corresponding group markers. 
Each element of the list contains the relative set of marker indexes as ordered in the original rowData(sim). 
The list doesn't keep the order of the ranking.
}
\description{
This function gives you the ranked list of group markers at the specified proportion of top markers (specificity)
}
\examples{

}
