% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore.r
\name{matchSCore}
\alias{matchSCore}
\title{Compute the cellular phenotype matching across simulated cell groups and computational clusters}
\usage{
matchSCore(gene_markers, gene_cl, labels)
}
\arguments{
\item{gene_markers}{A list of gene markers. Each element of the list represents the set of markers for each biological group.}

\item{gene_cl}{A list of cluster-specific gene sets to match with gene_markers.}

\item{labels}{A vector of gene_cl indexes correspondant to the gene_markers group level.}
}
\value{
A value from 0 to 1 representing the matchSCore between gene_cl and gene_markers. A matchSCore of 1 is the total match.
}
\description{
This function computes the matchSCore across gene markers provided from the simulation and 
cluster-specific genes computationally obtained from the tool.
}
\examples{

}
