% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore2-pkg.R
\docType{package}
\name{matchSCore2-pkg}
\alias{matchSCore2-pkg}
\title{matchSCore2}
\description{
\code{matchSCore2} is a Bioconductor package that provides ... # TODOELI, we need a nice short para for this
}
