% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore2.r
\name{matchSCore2}
\alias{matchSCore2}
\title{Compute the cellular phenotype matching across reference cell groups and computational clusters}
\usage{
matchSCore2(gene_cl.ref, gene_cl.obs, ylab, xlab)
}
\arguments{
\item{gene_cl.ref}{A list of reference group markers. Each element of the list represents the set of markers for each biological group.}

\item{gene_cl.obs}{A list of cluster-specific gene sets to match with gene_cl.ref.}

\item{ylab}{A character label (ylab) indicating the reference technology to plot in the matchSCore table(e.g. "Smart-Seq2").}

\item{xlab}{A character label (xlab) indicating the tested technology to plot in the matchSCore table (e.g. "Chromium").}
}
\value{
A list with the matchSCore matrix and the ggplot table at the ggplot slot.
}
\description{
This function computes the matchSCore across reference gene markers and 
cluster-specific genes computationally obtained from the tool.
}
\examples{

}
