% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_identity_heatmap.R
\name{ms_identity_heatmap}
\alias{ms_identity_heatmap}
\title{Heatmaps of class probabilities}
\usage{
ms_identity_heatmap(out)
}
\arguments{
\item{out}{output of the function \code{ms_identity_map}. # TODOELI: here as well, some param name more informative of what needs to be provided?}
}
\value{
A ggplot object, with the heatmap displaying the probability of each
cell to belong to the identity classes according the model.
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{
# TODO
}
