% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_summary_barplot.R
\name{ms_summary_barplot}
\alias{ms_summary_barplot}
\title{Cumulative barplot of cell mapped compositions}
\usage{
ms_summary_barplot(class.fac, obs.fac)
}
\arguments{
\item{class.fac}{A named vector of classified cells (e.g. the \code{output$ids}
from the \code{\link[=ms_identity_map]{ms_identity_map()}} function).}

\item{obs.fac}{A named vector of clusters.}
}
\value{
A cumulative barplot describing the cell identity compositions of the
observed clusters onto the reference
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{
# TODO
}
