% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_top_markers.R
\name{ms_top_markers}
\alias{ms_top_markers}
\title{Select the top n cluster markers}
\usage{
ms_top_markers(clusters, markers, ntop)
}
\arguments{
\item{clusters}{A vector of cluster labels.}

\item{markers}{A \code{data.frame} as in the output of the \code{FindAllMarkers} Seurat
function.}

\item{ntop}{The number of top markers you want.}
}
\value{
A list of ntop markers per cluster.
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{
# TODO
}
