% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_sim.r
\name{rank_sim}
\alias{rank_sim}
\title{Rank simulated group markers}
\usage{
rank_sim(sim)
}
\arguments{
\item{sim}{A Splatter sim object with simulated cellular groups or paths.}
}
\value{
A data frame containing the group gene ranking and other additional information for each gene from the simulated data set.
}
\description{
This function rank group markers according their relative specificity based on fold changes.
}
\examples{

}
