% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_scores_by_specificity.R
\name{tool_scores_by_specificity}
\alias{tool_scores_by_specificity}
\title{Compute matchSCore for a specified value of specificity and ntop ranked genes}
\usage{
tool_scores_by_specificity(sim, specificity, tool_out, ntop,
  tool_run = seurat_run, labels)
}
\arguments{
\item{sim}{A Splatter simulation object.}

\item{specificity}{The proportion of top-ranked markers for each simulated group.}

\item{tool_out}{The output of the used tool to cluster your data (It should be as the output of the seurat_run function).}

\item{ntop}{The proportion of top-ranked markers for each cluster.}

\item{tool_run}{The name of the function used to run the clusters without quote (Default= seurat_run).}

\item{labels}{Cluster labels as in the output of the compute_labels function.}
}
\value{
A matchSCore value
}
\description{
This function computes the matchSCore for different cutoffs of marker specificity and top cluster markers
}
\examples{

}
