% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscores_plots.R
\name{tscores_plots}
\alias{tscores_plots}
\title{Plots matchSCore trends for a specific tool}
\usage{
tscores_plots(scores_df)
}
\arguments{
\item{scores_df}{A character vector of reference group labels. TODO: df or vector?}
}
\value{
A plot object, TODO
}
\description{
This function identifies true label groups between reference groups and clusters
}
\examples{
# TODO
}
