% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_heatmap.R
\name{identity_heatmap}
\alias{identity_heatmap}
\title{Heatmaps of class probabilities}
\usage{
identity_heatmap(out)
}
\arguments{
\item{out}{output of the function identity_map.}
}
\value{
The heatmap displaying the probability of each cell to belong to the identity classes according the model.
}
\description{
This function identifies true label groups between reference groups and clusters.
}
\examples{

}
