% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore2.R
\name{matchSCore2}
\alias{matchSCore2}
\title{Compute the cellular phenotype matching across reference cell groups and
computational clusters}
\usage{
matchSCore2(gene_cl.ref, gene_cl.obs, xlab, ylab)
}
\arguments{
\item{gene_cl.ref}{A list of reference group markers. Each element of the
list represents the set of markers for each biological group.}

\item{gene_cl.obs}{A list of cluster-specific gene sets to match with \code{gene_cl.ref}.}

\item{xlab}{A character label, indicating the tested technology to plot
in the matchSCore table (e.g. "Chromium") on the x axis}

\item{ylab}{A character label, indicating the reference technology to plot
in the matchSCore table(e.g. "Smart-Seq2") on the y axis}
}
\value{
A list with the \code{matchSCore} matrix and the ggplot table at the
\code{ggplot} slot.
}
\description{
This function computes the matchSCore across reference gene markers and
cluster-specific genes computationally obtained from the tool.
}
\examples{
# TODO
}
