% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat3_run.R
\name{seurat3_run}
\alias{seurat3_run}
\title{This function runs Seurat3 by using the output of the \code{align_run} function.}
\usage{
seurat3_run(
  out_align,
  dims = c(1:10),
  res = 0.2,
  col_anno = NULL,
  col_data = NULL,
  verbose = TRUE
)
}
\arguments{
\item{out_align}{The output of the function \code{align_run}. The combined count
matrix is used to create the Seurat object and the integrated is provided
to the slot \verb{@data}.}

\item{dims}{Seurat parameter. It is the dimension of the PCA space.}

\item{res}{Seurat resolution parameter.}

\item{col_anno}{TODO}

\item{col_data}{TODO}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}
}
\value{
The integrated Seurat object. Two UMAP plots with cells coloured by
annotation and dataset will be generated.
}
\description{
This function runs Seurat3 by using the output of the \code{align_run} function.
}
\examples{
# TODO
}
