% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscores_plots.r
\name{tscores_plots}
\alias{tscores_plots}
\title{Plots matchSCore trends for a specific tool}
\usage{
tscores_plots(scores_df)
}
\arguments{
\item{scores}{A character vector of reference group labels.}
}
\description{
This function identifies true label groups between reference groups and clusters
}
